+++ Examples/cli/individual/EV_SB/getEvseDeviceState.js
@@ -1,0 +1,42 @@
+// NOTE: This command will only work on EV Smart Breaker Charger
+
+const {
+    EmcbUDPbroadcastMaster,
+    EMCB_UDP_EVENT_DEVICE_DISCOVERED
+} = require('../../../..'); // If running this example somewhere outside of a `git clone` of the `emcb-udp-master` module, replace with `require("emcb-udp-master")`
+
+const UDPKeys = require("../../../_config.js");
+
+var parsers = require('./evseExampleParsers.js');
+
+var EMCBs = new EmcbUDPbroadcastMaster({
+    broadcastUDPKey : UDPKeys.broadcast,
+    unicastUDPKeys  : UDPKeys.unicast
+})
+const argv = require('minimist')(process.argv.slice(2));
+var deviceID = argv._[0];
+var deviceIP = argv._[1];
+
+console.log("Establishing connection to device " + deviceID + " at ip: " + deviceIP);
+var device = EMCBs.createDevice(deviceID, deviceIP, true);
+
+EMCBs.on(EMCB_UDP_EVENT_DEVICE_DISCOVERED, (data) => {
+    console.log('Device connected from address: ' + data.device.ipAddress);
+    console.log("Sending: Get EVSE device state");
+
+    data.device.getEvseDeviceState()
+        .then(data => {
+            for (const [ip, response] of Object.entries(data.responses)) {
+                console.log("Response from " + ip + ":");
+                console.log(parsers.parseGetEvseDeviceStateResponse(response));
+            }
+        })
+        .catch(err => {
+            console.error("Failed to get EVSE device state")
+            console.error(err)
+        })
+        .then(() => {
+            process.exit()
+        })
+})
+